\name{anoctua}
\alias{anoctua}
\docType{data}
\title{
Presence-absence data of Little owls in nest boxes
}
\description{
The data contains presence-absence data of Little owls in nest boxes and elevation
}
\usage{data(anoctua)}
\format{
  A data frame with 361 observations on the following 3 variables.
  \describe{
    \item{\code{Id}}{nest box id}
    \item{\code{PA}}{indicator of Little owl presence}
    \item{\code{elevation}}{elevation (meters above sea level)}
  }
}
\references{
Gottschalk, T, Ekschmitt, K., Volters, V. (2011) Efficient placement of nest boxes
for the little owl (Athene noctua). The Journal of Raptor Research 45: 1-14
}
\examples{
data(anoctua)
}
\keyword{datasets}
