\name{blackstork}
\alias{blackstork}
\docType{data}
\title{
Breeding success of Black storks in Latvia
}
\description{
The data set contains number of nestlings in Blackstork (Ciconia nigra) nests.
}
\usage{data(blackstork)}
\format{
  A data frame with 1130 observations on the following 3 variables.
  \describe{
    \item{\code{nest}}{number of nest (nest id)}
    \item{\code{year}}{year}
    \item{\code{njuvs}}{number of nestlings}
  }
}
\source{
The data is property of Maris Stradz. Attention, this is a non-random subselection 
of the data available. Please, contact Maris Stradz, if you have interest in the 
whole data set.
mstrazds@latnet.lv
}
\examples{
data(blackstork)
}
\keyword{datasets}
