\name{blmeco-package}
\alias{blmeco-package}
\alias{blmeco}
\docType{package}
\title{
Data Sets and Functions Accompagning the Book Bayesian Data Analysis in Ecology Using Linear Models with R, BUGS and Stan
}
\description{
Data sets and functions accompagning the book Bayesian data analysis in ecology using linear models with R, BUGS and STAN
}
\details{
\tabular{ll}{
Package: \tab blmeco\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-03\cr
License: \tab GPL-2 \cr
}
See book
}
\author{
Fraenzi Korner-Nievergelt

Maintainer: Please, complain to <fraenzi.korner@oikostat.ch>
}
\references{
Korner-Nievergelt et al. book
}
\keyword{ package }
