\name{parusmajor}
\alias{parusmajor}
\docType{data}
\title{
Number of migrating Great tits
}
\description{
Counts of Great tits (Parus major) observed at the mountain pass Ulmethoechi 
(BL, Switzerland) between 1982 and 2007 during fall migration.
}
\usage{data(parusmajor)}
\format{
  A data frame with 434 observations on the following 3 variables.
  \describe{
    \item{\code{year}}{year}
    \item{\code{julian}}{day of the year}
    \item{\code{count}}{number of individuals counted}
  }
}


\references{
Korner-Nievergelt F, Korner-Nievergelt P, Baader E, Fischer L, Schaffner W, Kestenholz M (2007)
Jahres- und tageszeitliches Auftreten von Singvoegeln auf dem Herbstzug im Jura (Ulmethoechi,
Kanton Basel-Landschaft). Der Ornithologische Beobachter 104: 101-130.
}
\examples{
data(parusmajor)

}
\keyword{datasets}
