\name{periparusater}
\alias{periparusater}
\docType{data}
\title{
The data contain morphological measurements taken from museum skins of Coal tits (Periparus ater)
}
\description{
The data is part of the study by Korner-Nievergelt & Leisler (2004) Morphological 
convergence in conifer-dwelling passerines. Journal of Ornithology 145: 245-255.
}
\usage{data(periparusater)}
\format{
  A data frame with 28 observations on the following 6 variables.
  \describe{
    \item{\code{country}}{country of origin of the individual}
    \item{\code{age}}{numeric code for age categories as defined by www.euring.org, 
    3 = hatching year, 4 = not hatching year, 5 = after hatching year, 0 = missing}
    \item{\code{sex}}{numeric code for sex as defined by www.euring.org, 1 = male, 2 = female, 0 = missing}
    \item{\code{weight}}{body mass in g}
    \item{\code{P8}}{length of primary 8 in mm. Primary 8 is the third outermost wing feather often building the wing tip.}
    \item{\code{wing}}{wing length in mm}
  }
}

\references{
Korner-Nievergelt & Leisler (2004) Morphological 
convergence in conifer-dwelling passerines. Journal of Ornithology 145: 245-255.
}
\examples{
data(periparusater)

}
\keyword{datasets}
