\name{redstart}
\alias{redstart}
\docType{data}
\title{
Common Redstart (Phoenicurus phoenicurus) counts
}
\description{
Counts of Common Redstart (Phoenicurus phoenicurus) breeding pairs between 1993-1996 in a small part of Switzerland.
}
\usage{data(redstart)}
\format{
  Data frame with 342 observations and the following 5 columns:
  \describe{
    \item{\code{counts}}{count of Common Redstart breeding pairs in each 1 km2 plot}
    \item{\code{x}}{x-coordinate in CH1903-LV03 (EPSG: 21781)}
    \item{\code{y}}{y-coordinate in CH1903-LV03 (EPSG: 21781)}
    \item{\code{elevation}}{average elevation in m.}
    \item{\code{forests}}{forest cover}
  }
}
\source{
Swiss Breeding Bird Atlas 1993-1996 (Swiss Ornithological Institute): http://www.vogelwarte.ch
}
\references{
Schmid H., Luder R., Naef-Daenzer B., Graf R., Zbinden N. (1998) Schweizer Brutvogelatlas. Verbreitung der Brutvoegel in der Schweiz und im Fuerstentum Liechstenchstein 1993-1996. Schweizerische Vogelwarte, Sempach.
}
\examples{
data(redstart)
}
\keyword{datasets}
