\name{resprouts}
\alias{resprouts}
\docType{data}
\title{
Survival data of tree sprouts
}
\description{
Number of tree sprouts that survived a management fire and the time since the last fire.
}
\usage{data(resprouts)}
\format{
  A data frame with 41 observations on the following 4 variables.
  \describe{
    \item{\code{treatment}}{time since last fire in months}
    \item{\code{plot_ID}}{plot name}
    \item{\code{pre}}{number of tree sprouts before the fire}
    \item{\code{post}}{number of tree sprouts after the fire, survivors}
  }
}
\references{
Walters, G (2012) Customary fire regimes and vegetation structurein Gabon's Bateke Plateaux.
Human Ecology 40: 943-955
}
\examples{
data(resprouts)
}
\keyword{datasets}
