\name{roostingsiteuse}
\alias{roostingsiteuse}
\docType{data}
\title{
Roosting site use by little owls
}
\description{
Locations of roosting sites of little owls obtained by telemetry data
}
\usage{data(roostingsiteuse)}
\format{
  A data frame with 42 observations on the following 5 variables.
  \describe{
    \item{\code{roosting.loc}}{a factor with 4 levels}
    \item{\code{roostingnum}}{roosting site number}
    \item{\code{temp}}{ambient temperature in degree celsius}
    \item{\code{familynum}}{number of the family}
    \item{\code{indnum}}{number of the individual}
  }
}
\references{
Bock, A., Naef-Daenzer, B., Keil, H., Korner-Nievergelt, F., Perrig, M., Grueebler, M. U. 
(2013) Roost site selection by Little Owls Athene noctua in relation to environmental 
conditions and life history stages. Ibis 155: 847-856.
}

\examples{
data(roostingsiteuse)
}
\keyword{datasets}
