\name{wildflowerfields}
\alias{wildflowerfields}
\docType{data}
\title{
Territory numbers of Whitethroat in wildflowerfields
}
\description{
Number of territories of Whitethroat in wildflowerfields of different ages. The data
has been collected by J-L Zollinger.
}
\usage{data(wildflowerfields)}
\format{
  A data frame with 136 observations on the following 8 variables.
  \describe{
    \item{\code{field}}{field id}
    \item{\code{year}}{year}
    \item{\code{age}}{age of the wildflower field in years}
    \item{\code{bp}}{number of territories of whitethroats Sylvia communis}
    \item{\code{X}}{x-coordinate}
    \item{\code{Y}}{y-coordinate}
    \item{\code{size}}{area of the field in ares (a, 10 x 10 m)}
    \item{\code{Nspec}}{number of species}
    
  }
}
\references{
Zollinger J-L, Birrer S, Zbinden N, Korner-Nievergelt F (2013) The optimal age
of sown field margins for breeding farmland birds. Ibis 155: 779-791
}
\examples{
data(wildflowerfields)
}
\keyword{datasets}
